- Fix issue with decimal points after recent Nako indexer upgrades. - Adds a custom tag helper to display coin, making it possible to individually style it.

+++ Stratis.Guru/Models/ApiModels/TransactionModel.cs
@@ -19,7 +19,7 @@ namespace Stratis.Guru.Models.ApiModels
 
         public string CoinBase { get; set; }
 
-        public decimal Value { get; set; }
+        public long Value { get; set; }
 
         public long BlockIndex { get; set; }
